+++ components/Select.js
@@ -5,18 +5,7 @@ import styles from '../styles/errors.module.css'
 
 export default function Select(props) {
   const [field, meta] = useField(props); // implement usefield
-  const [dropDownOption, setDropDownOption] = useState([]);
-
-  const genderOption = [
-    {
-      id: 1,
-      name: "Male",
-    },
-    {
-      id: 2,
-      name: "Female",
-    },
-  ];
+  const { selectOptions, selectValue } = props;
 
   // for region, municapality, district options
   // useEffect(() => {
@@ -31,20 +20,30 @@ export default function Select(props) {
   //   fetchOptionData();
   // }, []);
 
-  const Options = field.name === "gender" ? genderOption : dropDownOption;
-
   return (
     <div class="my-3 is-flex is-flex-direction-column is-align-items-flex-start">
-      <label htmlFor={field.name} class="label has-text-left is-uppercase">{field.name}</label>
-      <div class={`select is-large is-capitalized is-primary ${meta.touched && meta.error && 'is-danger'}`}>
-        <select class="is-capitalized is-uppercase" {...field} {...props}>
-          <option value="" disabled selected>select {field.name}</option>
-          {Options.map((option) => (
-            <option>{option.name}</option>
+      <label htmlFor={field.name} class="label has-text-left is-uppercase">
+        {field.name}
+      </label>
+      <div
+        class={`select is-large is-capitalized is-primary ${
+          meta.touched && meta.error && "is-danger"
+        }`}
+      >
+        <select class="is-capitalized is-uppercase" {...field} {...props} value={selectValue ? selectValue : null}>
+          <option value="" disabled selected>
+            select {field.name}
+          </option>
+          {selectOptions.map((option) => (
+            <option key={option.id}>{option.name}</option>
           ))}
         </select>
       </div>
-      <ErrorMessage component="div" name={field.name} className={styles.error}/>
+      <ErrorMessage
+        component="div"
+        name={field.name}
+        className={styles.error}
+      />
     </div>
   );
 }